meta-java: bcel,avalon-framework: Specify encoding to javac
when using javac from different JDK it does not like
the encodings since default is ascii, so be specific.

fastjar argument order was not correct when using other
fastjar implementation it was strict about it so lets
correct it as well.

(From meta-java rev: b81f7aafb3f8047faf89b1dd089062bcf2035462)

Signed-off-by: Khem Raj <raj.khem@gmail.com>