meta-openembedded: ckermit: Explicitly set EXTRA_OEMAKE as required
This recipe currently relies on EXTRA_OEMAKE having been to set to "-e
MAKEFLAGS=" in bitbake.conf to operate. It is necessary to make this
explicit so that the default in bitbake.conf can be changed.

(From meta-openembedded rev: f941c1b36418f5436b76d0f687d93df597a17c82)

Signed-off-by: Mike Crowe <mac@mcrowe.com>
Signed-off-by: Martin Jansa <Martin.Jansa@gmail.com>