meta-openembedded: nbench-byte: Explicitly set EXTRA_OEMAKE as required
This recipe currently relies on EXTRA_OEMAKE having been to set to "-e
MAKEFLAGS=" in bitbake.conf to operate. It is necessary to make this
explicit so that the default in bitbake.conf can be changed.

(From meta-openembedded rev: 2cbf2ac87c8f6f92d4ec818678f1cd072fa9d996)

Signed-off-by: Mike Crowe <mac@mcrowe.com>
Signed-off-by: Martin Jansa <Martin.Jansa@gmail.com>