meta-intel-iot-security: xattr-images.bbclass: Set the floor label explicitly
On a target device where SMACK is active each new file gets assigned a label,
typically the one from the process creating the file. If a file that is
about to be installed or updated as a part of a swupd update doesn't
have any label then the default label assigned by the kernel on
the target device will make the hash of the file on the target device
differ from the hash registered in the update's Manifest file.

In order to avoid such discrepancy all files need to be labeled
explicitly hence set "_" explicitly on everything that would not
have a label otherwise.

(From meta-intel-iot-security rev: 00db2f843c31b502fb9fc2f339e0d20835074589)

Signed-off-by: Dmitry Rozhkov <dmitry.rozhkov@linux.intel.com>