dkpg: Use tar everywhere (not gtar)
All our systems usually have tar, if we want the native sstate to work
universally, we need to prefer this. This avoids a system with gtar
causing dpkg-native to use it and it not being present on some systems.

(From OE-core rev: d683913119082f718af64f2d402bac67b660fca6)

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>