ostro-os: remove meta-ostro-fixes
Applying changes to upstream components in a separate repo makes it
difficult to integrate new upstream components if there are
conflicts.

For example, for https://github.com/ostroproject/ostro-os/pull/3
several meta-ostro-fixes changes have to be removed before integrating
new upstream components, but removing the fixes first breaks testing
because it (temporarily) leads to a state where crucial fixes are
missing.

In addition, maintaing changes as .bbappends is cumbersome and does
not work in all cases (for example, modifying upstream .bbclass files
or data files).

Therefore it is better to do maintain upstream changes as patches in
ostro-os, which is where these fixes will be moved to.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>