meta-appfw: ostro-app.bbclass: change permissions after relocation.
Change application permissions after relocation to prevent
reading or execution of files across applications (as
requested by the security folks). We do this by changing
ownership to root, group ownership to the application and
clearing all permissions from 'other'. We also propagate
set the group readability and executability bits to make
sure the application can access its own data.

This is probably incomplete and will need more eyeballing.
What we should eventually do is to set everything write-
protected and readable only by the app. In the manifest
the applications should be able to request a writable
directory for storing its own data/state. We cannot do that
until the filesystem layout is properly finalised and
documented. This should include things like how the layout
is if/when/once we have a readonly root, where apps can store
persistent data that is not supposed to be lost across box
resets (if such a thing is supported), etc.

Meanwhile we try to go with this.

(From meta-appfw rev: cefa118055c7755b2d66c60b7ef52e75c5f1cb32)

Signed-off-by: Krisztian Litkey <krisztian.litkey@intel.com>