ostro-image.bbclass: remove systemd-firstboot.service
The service makes no sense in Ostro OS because it relies on asking
questions during the initial boot, at least as it is configured at the
moment. We could use it to set pre-defined values, but then we can as
well just set them directly in the images.

Not having systemd-firstboot.service allows removing /etc/machine-id,
which is important both for stateless and the intermediate solution
with mixed maintenance of /etc (some files updated via swupd, others
not).

Work is underway to make systemd more modular; in the meantime, the
approach used here with removal after installation already works and
will not cause problems either when systemd-firstboot does not get
installed anymore.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>