swupd: enabled updating /etc
As an intermediate step towards full stateless Ostro OS, we now
treat some files in /etc as conceptually read-only (i.e. neither
modified by the OS at runtime nor by an admin). Anything contained
in the rootfs directories will get bundled and added or updated when
running "swupd update".

The implication is that we must keep certain files out of the rootfs
which do get modified at runtime, because otherwise there are
"swupd verify" failures.

"/etc/machine-id" needs special treatment: it has to be in images, but
must not be in bundles. Uses the approach currently being tested by
Joshua for meta-swupd.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>