meta-ostro: supported-recipes.bbclass: skip dependency section if empty
When the unsupported recipe is directly one of the build targets (as
in "bitbake foo"), there is nothing that depends on it and the
dependency chain section becomes empty. In that case it can and should
be skipped entirely.

(From meta-ostro rev: 1e7bb4da72d4f1e5d8d88dfd5554292da307b59f)

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>