meta-ostro: ostro os: split up supported recipes list
The @workspacelayer may be useful also without the list of recipes
supported by the Ostro project, therefore it is better to have it in a
separate file.

Also demonstrates that SUPPORTED_RECIPES is a list of file names.

(From meta-ostro rev: 406ee6f080bab31e0acde2557ea9a1d11ce32900)

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>