meta-ostro: supported-recipes.bbclass: move native special cases into variables
The main advantage is that it is easier to read and explain this
way. The expectation is that these lists will get maintained together
with and in the class itself, but should it become necessary to override
these special cases, then that is also possible now.

(From meta-ostro rev: 1281d85fd790a033ad73b64a53949456ece748c5)

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>