meta-ostro: supported recipe: explain why a recipe is supported
When listing a recipe as "supported", it is useful to know who
supports it (i.e. which distro or person). This is derived from the
file(s) in which the recipe was listed.

The revised report adds a space-separated list of file base names, or
if set, the corresponding value of the SUPPORTED_RECIPES varflag in
round brackers after "yes". Example: "supported: yes (ostro)"

(From meta-ostro rev: 0c6c69f2717035dc4050986a1f00f336b0cda845)

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>