meta-security-isafw: ISA_cfa_plugin.py: explicitly treat all command output as utf-8
Python3 returns output as byte strings. We need to decode that first
before we can work with the output as normal strings. For example,
without this change the following error occurs for
'"symlink" in file_type':
  TypeError: Type str doesn't support the buffer API

Assuming that the output is utf-8 encoded is a simplification,
but should be a safe bet (also works for plain ASCII).

(From meta-security-isafw rev: d5ec01ba4aab1c672ad694cd95e045f456d0fed8)

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>