test-iot.bbclass: Python3 compatibility
The loader mechanism dropped the "filepath" attribute. Directly
using the known path via COREBASE is easier.

Python3 returns generators from map() and filter(). When
iterating multiple times, we need to copy into lists because
otherwise iteration only works the first time. When checking
for a match, we cannot test the filter() result, because
it is always a valid object and thus evaluates to "True". Instead,
check each element in turn (map) and stop checking when any
tests true (any) - this is (in theory) also a bit faster under
Python2.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>