meta-edison-bsp: u-boot: fix target_env archiving
With the recent archiver.bbclass changes in oe-core an error
in do_ar_original was triggered when archiving u-boot.

It turns out archiver detects 'file://target_env/' as another
source URL and wants a unique repo name for it (or the original
source).

To avoid the need to change the archiver.bbclass logic, just give
a unique name to target_env "repo" (directory).

While debugging this, it was noticed the old archiver.bbclass was
not archiving file://target_env/ at all. This change (together with
the new archiver.bbclass) fixes that too.

Signed-off-by: Mikko Ylinen <mikko.ylinen@intel.com>